Symbolic differentiation.
Added SD.scala, a module which applies symbolic differentiation.
The main functionality is there, but quite a lot of cases in the
AST have not been handled. There are TODOs for these in the code
but a selection of them is:
 - Fields / private variables of a ClassDef
 - The test of an if statement
 - The "subject" of a switch statement
 - LHSs of equations

Added SD to the AST transformation pipeline in AppModel.scala.