Enable proguard.
It works, but needs some fine tuning so it creates a jar file that is
similar to the sbt 0.7.7 version.

The path of least resistance was to fetch the source from git and have
sbt compile it.  Unfortunately, the official git repo. doesn't compile
with sbt 0.12, so I had to use an unofficial version with the fix in.
Just to be on the safe side the build depends on a local fork (in
the github sense).