Added comment and helpful variable names in contractDomain.
The idea behind HC4 is that expressions can be evaluated in both
directions, rather than just from variables to root. The reverse
direction starts with an interval that is the sought range of the
expression, which is then propagated backwards through the expres-
sion using the so called inverse relations of the operations that
the expression is build up from. For example, if e = x+y, then the
two reverse relations for e are x = e-y and y = e-x. The intervals
obtained by back propagation are intersected with the intervals
for x and y to obtain a contracted domain for e that "corresponds
with" the constraint that the range of e should lie in the propa-
gated interval.