Added more general predicate support computation.
The support is the set of values in the domain for which the given
predicate is true. The support computation takes a predicate and
a domain box and shrinks the box as much as possible while still
containing all of the part of the support of the predicate that
was included in the box.