Rewrite AppModel.  Pause and Step now work for all Interpreters, including the Enclosure interpreter.
AppModel is renamed to "Controller" since that more accurately
describes what it is doing: controlling the interpreter models.
Instead of having the AppModel spawn a consumer and producer actor,
the new Controller class _is_ the consumer actor and the producer part
is moved out into a separate InterpreterModel class.  This class, so
named because it loosely models the interpreter, has one main method,
"produce", which spawn a new producer actor for the interpreter its is
modeling.  There are two different models.  One for the CStore data
and one for the new Enclosure interpreter.

In addition, a new class AppState is created to represent the main
state of the application in an attempt to simplify state handling.

Some Acumen3D tests are broken as they no longer compile and need to
eventually be fixed.  For now they are commented out and an exception
is thrown in its place.