Refactor Interpreter trait.
The two interpreter types (CStore and Enclosure) really have nothing
in common so don't pretended they do and create a separate interpreter
trait for the Recursive (Enclosure) interpreter.  This is okay now as
the Controller (was AppModel) no longer directly interfaces with the
interpreters and instead uses the new InterpreterModel.

This broke running the Enclosure interpreter form the command line,
but I don't think that really worked anyway.