Don't let plotting tasks pile up.
That is, when it is time to produce the next plot and multiple plotting
tasks are queued, ignore all but the latest task.

This is as oppose to forcing the interpreter and plotter to be in
sync, which was the previous hack.  Previously if the plotter took longer
than the interpreter to plot a given chunk, the interpreter was forced
to wait.