Kill TraceModel as it is currently known and other misc changes.
TraceModel was the model for both the Plot and Trace view, then the
interpreter model become the real model for the data and TraceModel
was just a proxy.  Now, eliminate the need for a proxy and make the
TraceModel a model for just the Trace view and have a separate model,
PlotModel, for the plot view.  The interpreter model now instead has
methods for returning new instances of the Trace and Plot model
(getTraceModel/getPlotModel).

For efficiency reasons and to minimize possible data contention the
addData method in the Interpreter model does not update the models,
instead it caches pending changes and then flushes the data once a
get*Model method is called.

Also, have the trace view listen to app. state changes and respond
accordingly.  For now do not update the trace view every time data is
added, instead have the trace view update itself when the state
changes to "Ready".

All these changes are prep work for having the plotter on a separate
thread.