Move CPU intensive plotter code to a separate thread.
Most features, such as zooming and saving an image work.  The only
feature that doesn't work is the "Draw" check box.  Save works but it
does all its work on the EDT and should eventually be fixed.

To allow the plotter to be on a separate thread, a new Plotter actor
is now responsible for the plotting.  Whenever the plot needs to be
redrawn (either due to updated data or a new view on plot from zooming
or resizing the window) a message is sent to the plotter actor.  The
plotter responds by getting the data it needs from the
InterpreterModel and when it's down, it updates PlotPanel with the new
plot image and calls its repaint method to update the display.  To
keep plotting tasks from piling up duplicate messages are discarded.

The plotter, as a matter of convince, also updates the TraceView
Model.  The actual plotting is broken down into several phases with
only the last phase now on the EDT.  First the Graphics2D objects are
created, then the Plot is drawn on a canvas (really a buffer), and,
finally, and on the EDT, the updated image (the buffer) is drawn on
the display.  Any of the earlier phases can be skipped if they are not
necessary.  For example, when zooming the Graphics2D objects do not
need to recreated.

To make the phases cleaner, and to simplify the code the paint method
in PlotData is now a class (PlotImage).  This class, like PlotData
does all its work when initialized.