Fix "Draw" checkbox and related refactoring.
Group all input for the plotter into the class PlotInput so that it is
all in one place.  In addition, instead of having the plotter actor
directly updating the Plot and Trace view when data is ready, it sends
a message to the Acumen EDT actor with all the necessary data.  This
message/event is picked up by the Plot and Trace view, which then
update themselves as appropriate.

To fix the "Draw" checkbox a volatile state variable, that is part of
PlotInput, is used.  This value is simply changed by the checkbox.
The plotter and PlotPanel (Plot view) then check for this variable and
act appropriately.