Cleaned up some deprecated files and patched intersection bug.
The intersection with domain invariant taken when computing the
initial condition for a given event sequence before solving the
ODE during event tree enlargement could be empty. When this happens
an exception is thrown, but it may not signify an inconsistent
model, but can mean an impossible event. Right now the exception is
caught and the event excluded from the event sequence. Have to
re-visit this and make sure we discover when this happening really
flags ill posed models.