More Tweaks in App. level events and better behavior for when the plot zoom etc buttons are active.
The Plot panel now reports its status as to if it is enabled are not.

The Plotter now just uses the status as the event message instead
of wrapping it in PlotStateChanged.

Finally, the new Plot panel events are used to define more sensible
behavior for enabling and disabling of plot zoom etc buttons.

In addition, the Acumen object now has a "publish" and "!" shortcut
methods.