Partly revert "Implemented ticks for enclosure semantics plotter."
This is in prep for merge with the refactor-ui branch, will reapply
if the functionally is needed.

This reverts commit 5fc80d1d6c9c203a886fac254a08ce69d7d134e5, except
that the file "src/test/scala/acumen/Sandbox.scala" was not reverted
due to a conflict that wasn't worth resolving, as it looks like an
unrelated change that got committed at the same time.