Added basic PDF export functionality.
PDFs are now exported using separate JPopupMenu item.
This should be changed so that PDF is added as an option in the
Save As.. menu item dialog instead.

Also, currently there is a lot of code duplication with three
copies of the SaveFileDialog class in three different places.
Made a separate file out of this class to prepare for a later
refactoring, so that this class can be used both by the JPopupMenu
and by the Save button in the GUI.