Fixed bug preventing the solver from discarding impossible events.
When extending an event sequence an impossible event may still be
considered, but during enclosure computation it may be discovered
that no solution to the system exists as a consequence. Then the
solver treats the preceding event as possibly last and skips to
the next possible extension.

Also fixed a copy-paste type in the initialisation of the
minImprovement parameter in most exampe models.