Added minImprovement and splittingDegree simulator parameters.
minImprovement makes it possible to set the amount by which sub-
division of time segements in hte top level strategy needs to
improve the initial condition for the following time segment to
warrant more subdivision. Currently the measure used to compare
initial conditions is the L1 norm of the box, i.e. the sum of the
widths of the initial condition intervals. The choice of this
measure was somewhat arbitrary, and should be revised as it can
affect the performance of the startegy.

splittingDegree is a parameter that decides what number of pieces
each initial condition interval shall be split into when solving
IVPs. Splitting is a first attempt at getting thinning enclosures
for stable systems, such as x'=-x or x''=-x'-x.