Move iTextPDF dependency into XXX_build.sbt due to licensing issues.
The idea is that we we go to make a release all XXX files will be
excluded.  This is turn will cause the build to fail (due to a missing
dependency) and remind us to comment out (or do something) with the
code that depends on it (Export to PDF).