Added Pi constant to syntax and interpreters.
The reference and parallel interpreters use the Pi value provided
by the scala.math package. The enclosure interpreter uses a
hard-coded value based on a string representation of the 1000 first
digits of Pi. An interval is built out of the full precision
BigDecimal, based on the precision of the computation in which
the Pi constant is used. Currently, attempting to compute Pi with
a higher precision leads to a runtime exception.