Add -dontshrink to the list of proguard options.
Proguard doesn't always get the shrinking right and can remove fields
that are accessed elsewhere.

One particular bad example is when accessing the PLOT_IDX field in an
anonymous class (that extends TabbedPane) created in ui.App.  Proguard
removed this field and an exception was thrown when the field was
accessed inside PlotTab, yet the exception was silently caught and
ignored (the code was executed on the EDT); thus, leading to a rather
difficult to find bug.  (Relevant commits: d13538ce (Disable plotting
when ...) and 7a3a8c60 (Disable trace table updates when ...)).