Move examples from src/main/resources/acumen/examples/ to just examples/.
As the examples are no longer a resource, this avoids the danger of
saving a modified example to the wrong location.

Have Acumen's file picker look for "examples" in the current dir and
if it's there, use that, otherwise just use the current dir.

Fix "sbt test" to work with the new location.  It is assumed that it
is run from the git root dir.

This reverts commit 4b7718e2 and 23a54fb6 which added the "--home"
option as it no longer needed, and also doesn't work reliably.