Added first working version of the localizing strategy.
The new strategy uses two separate parameters to control stepsize
in the solver. minTimeStepODEsolving is the minimum time step used
when IVP solving while minTimeStepLocalising gives the time step
used by the localization algorithm.

When debugging the solver a bug in restrictTo was also found and
fixed in UnivariateAffineEnclosure and UnivariateAffineScalarE..

The incremental plotting should be fairly straightforward to
implement with the current form of the localizing strategy but the
currently commented out callback hooks are not placed correcly, or
should filter some of the enclosures passed to them in
computeLFEnoODE.