Prepared and commented out parts of Picard overflow handling.
Application of the Picard operator can lead to numeric overflow at
times, such as for the current version of the bouncing_ball_air
model. The likely cause is that the time segment over which the
operator is applied is so large that the operator is no longer a
contraction, leading to a numerically divergent computation.

A likely fix is to use exception handling of some sort (possibly
by making solveVt return a maybe enclosure) to guide splitting of
the time segment. This spliting can be done external to solveVt,
or internal, producing a sequence of enclosures. These can then
either be passed, or if one prefers to retain the single enclosure
return type, stitching of the enclosures into a single one can be
performed. This operation is likely to be fairly expensive, but
may also produce an enclosure with an overall smaller amount of
splitting. Note that generalising the return type to sequences of
enclosures opens up the possibility to representing the nothing
option of the maybe enclosure as the empty sequence.