Removed unnecessary and costly "evaluation at corners".
The removed algorithm evaluated the enclosure at the corners of
the domain, which is exponentially costly, but was used as tempo-
rary solution that was supposed to give the most precise evalu-
ation possible to remove some of the sources of hard-to-discover
infomration loss from the development process. As it turns out the
extra effort is never warranted for affine enclosures.