Fixed CStore interpreter tests.
The fact that InterpreterTestBase.testExamples was running tests
for all models in all recursive subdirectories of the examples
directory was causing the CStore interpreters tests to fail.
This is due to the fact that, currently these interpreters do not
support the simulator parameters of the enclosure interpreter.
A TODO has been added as a reminder to add support for these
parameters to the CStore interpreters (also see ACUMEN-75).