Don't let JFreeChart slow down the Application.
The key part is allowing the new plotter (JFreeChart) to be disabled and
reenabled again.  The new plotter is now only enabled when the tab is
selected (like the old plotter) and can also be disabled by the user.

Also, since live updates have to take place on the EDT, disable live
plotting when the updates are taking too long.  Instead once the
simulation is done, detach the plot, create it in the background and
than attach it, which minimizing the amount of work that has to be
done in the background.