Disabled syntax highlighting for files other than Acumen models.
Also refactored the code which enables and disables the code editor
into separate functions, so that this could be resued (only used
for enableEditing currently).

This is a fix for ACUMEN-64 until we have decided whether or not to
go for some kind of markup format support for our README files.