Added a View menu to the menu bar (font config. and line numbers).
The View menu includes items for configuring the code editor:
 - Increase font size
 - Decrease font size
 - Reset font size
 - Font
 - Show line numbers

The Font menu item shows monospace fonts that are available to use
on the system. The total set of available fonts is intersected with
a list of known monospace fonts which covers standard fonts for
Mac OS, Linux and Windows, as well as some common programming
fonts. The default monospace font provided by the JVM is also
listed (i.e. the logical font corresponding to the name
"Monospace").
The "Reset font size" menu item sets the font size to 12, which is
also the default used when Acumen is started.

[Towards ACUMEN-84 ACUMEN-85]