Streamlined keyboard shortcuts.
Added mnemonics (to navigate menus using keyboard) to all menu
items and accelerators (direct keyboard shortcuts, with hint text
next to the corresponding the menu) for the most common menu items.

A couple of keyboard shortcuts were left as application-wide
shortcuts:

 - Ctrl + G   ~ Run simulation (Play)
 - Ctrl + T   ~ Stop simulation
 - Ctrl + B   ~ Step simulation
 - Ctrl + =   ~ Increase font size

The last one was added as a complement to Ctrl + "+", as this is
standard in most editors, due to the fact that "=" occupies the
same key as "+" on many keyboard layouts.

The three simulation-related kaybord shortcuts outght to be turned
into regular menu items with accelerators under a "Simulation"
menu, to make separate documentation of these shortcuts less
pressing.