Create script to remove Non-BSD code and mark code for such script.
That is all Non-BSD compatible code is either moved in an XXX file or
marked with a "// NRL" comment, where NRL = Non Release Line.

The script censor.pl removes all XXX files and lines marked with NRL.
For safety the script will refuse to run unless the file
"READY_FOR_CENSOR" is created.