Fixed thread leakage problem in both parallel inerpreters.
Because new instances of the interpreters were being created each
time the user selected how many threads to use, threads were being
retained. The fix for this was to maintain a singleton interpreter
and to set the scheduler (and its thread pool) on this object to
avoid creation of new instances.

Threads in the two thread pools (Static and Sharing) are now pre-
allocated when the interpreter is first selected (when its class
is loaded). The user can select any number of threads to be used
during simulation as long as it is lower than the number of
available cores, otherwise this number defaults to the number of
available cores (and the user is informed through a console
message).

This code still needs to be tested, particularly with respect to
performance on different classes of models, as the work sharing
scheduler is likely to perform bad when the amount of work done
in each object is small compared to the number of objects as well
as the amount of communication between them.