Refactored sequential imperative interpreter to use Common library.
Moved the sequential and parallel imperative interpreters into
their own packages under the imperative package.

There is still room for abstraction here as the implementations of
the step function for both of the imperative interpreters is
identical up to the call to iterateSimple/iterateMain. This should
consequently be passed to a shared implementation of step.