Exposed scheduler as a CLI paramter (to the --semantics argument).
This was done in order to be able to compare different schedulers
through benchmarking.

Now the CLI argument --semantics takes two new arguments:

 - parallel-static  The original parallel interpreter described in
                    the POOSC paper. Statically assigns work to
                    threads in the beginning of the traversal.
 - parallel-sharing A scheduler which dynamically reassings work
                    to threads using a work queue.

Fixed bug in the reset method of the SharingThreadPool which
caused it not to disable threads when resetting the number of
active threads to a number smaller than the number of cores
available to Acumen. Also refactored AcumenThread API in this
thread pool.

Now there are two semantics