Add a basic Acumen to C compiler for single object models.
The compiler generates C code (in model.c) that dumps the results in
tabular form that is similar to the table presented in the GUI.  Both
discrete and continuous steps produce data.  In addition a very basic
gnuplot script is outputted (in model.gpl) for visualizing the results.

For basic usage info see XXX_misc/scripts/testCompiler.sh and
examples/XXX_internal/compiler/ for test models.

Known limitations:

  - Only works on a single object model.
  - A few functions are unimplemented
  - Can not handle vectors
  - Assumes the all int constants are doubles
  - Can not handle dynamic types, assumes the type of a var. will not
    change after it is initialized.
  - Rarely used constructs (such as lists) are unimplemented.