Initial implementation of exposing the CStore through a socket interface.
This adds a server option to acumen.Main. It takes the port number
to listen on as an argument. An example invocation parameters are:
  /path/to/script.acm listen 1234
The server then blocks until a client connects to this port, after
which the simulation is performed and each simulation step's CStore
is sent through the socket to the client in a JSON format. The client
replies with a possibly modified CStore in JSON format, and this
data is used to perform the simulation step in Acumen.