Hack to be use command line main from within the graphical one.
Since the command line syntax for the two entry points are not the same,
detect which syntax is used and call the appropriate version.

This means that there is no longer any need for "run-main" with sbt or
building a special jar file in order to use the command line version
of acumen  (Although for now it is still possible to build a command
line only version.)

Towards ACUMEN-80.