Removed legacy plumbing code (solveVtE).
This code was used to provide plumbing to make it possible to
drop-in EncloseEvents in place of SolveVtE in the Solver trait
(by providing an implementation of the solveVtE function).

With these changes, the Event Tree based solver is functioning and
selectable from the Semantics menu of the Acumen IDE.

This fixes ACUMEN-141.