Added stub of PaddedUnivariateAffineEnclosure.
Main implementation issue is to figure out how to pad the parent
UnivariateAffineEnclosure with the appropriate bound.

Explanation: the end result should be such that that when the pad-
ded enclosure was constructed by C^1-Lohner theorem 5 padding of
a solution starting at the midpoint of the initial condition box,
then the L^1 norm of the padded enclosure contains each solution
starting from the initial condition box pointwise.

To avoid excessive padding we may have to compute the "padding
weights" on a component basis.