Add a basic type checker for the Acumen language.
The type checker currently does basic type inference via a
whole-program analysis and thus does not require any additional type
annotations by the user.

The type system relies primary on the type used to initialize the
variables, and as such if an Acumen class is unused the type inference
will be incomplete since the type of the constructor parameters will
not be known.  After initialization, the type can be adjusted
depending on how it is used.  For example, when a variable is
initialized with an integer constant the variable is assumed to be an
integer, but will be adjusted to a NumericType (ie double, ie real) if
it gets assigned to one.

The type checker tries to statically determine the type.  When that is
not possible a special DynamicType is used, for example if a variable
is used as both a Boolean and a string.

Right now the type checker is not used unless explicitly called from
the command line.  The command line version will output any type
errors or unification failures that result in a dynamic type being
used, a summary of the problems encountered, and a pretty printed
version of the program with type information on all class fields and
constructor parameters.

Towards ACUMEN-143.