Added first prototype of Lohner solver.
This version uses the logarithmic norm of the Jacobian of the IVP
field to derive a bound on the difference between solutions star-
ting in a given initial condition box. If the norm is negative over
the whole box, then the difference is bounded by the norm of the
box scaled by an inverse exponential fucntion of time. As a result
the solution stands a good chance of narrowing.

This marks progress on issue ACUMEN-156.

The current issues are:

1) The implementation of the interval exponential function needs
a serious review. It is probably not corerct for all precisions and
also needs cleaning up. Some of its components, such as the func-
tions for scaling of arguments, needs factoring out as they will
likely be useful in other places.

2) there is a memory leak causing the solver to grind to a halt for
non-1-dimensional systems. This is likely caused by a very naive
symbolic differentiation implementation of Jacobian computation.

3) The solver hooks in to the pwl hybrid algorithm by producing
enclosures that inherit from the enclosure type that is currently
hard-wired into most of the enclosure interpreter. The wrapping and
unwrapping of enclosures is likely causing garbage collection
overheads, also it perpetuates the tight coupling currently making
experimentation with the solver cumbersome.