Filter out simulator parameters specific to only one interpreter when used with another.
Parameters not known by any interpreter are left alone so misspellings
will still lead to an error.  Also, only assigns to the simulator
parameters are filtered; other uses of the simulator parameter are
not.

During startup the interpreters register the parameters they use and
the filtering is done via the new CleanParameters pass that the
interpreters call during initialization.

Towards ACUMEN-75.