Added first prototype of Interval cosine.
The implementation relies on a cached "enclosure table" to give
approximations for values for arguments in the range [0,pi/2].

It will act as a placeholder for a more refined implementation
while modularising and integrating with the enclosure interpre-
ter. Also it is an interesting proof-of-concept for the idea
of pre-computing enclosures that are likely to be used multiple
times and that of using the IVP solving facilities for obtaning
special functions.

The implementation is limited in some ways, in particular it
will require small minSolverStep and MinComputationImprovement to
produce reasonably precise results. More importantly, it uses
naive algorithms for shifting the argument interval into succes-
sively smaller ranges until the one for which the enclosure table
is available, causing an unnecessarily large amount of wrapping.

The initial use of this and related functions will be to provide
interval elementary functions for producing

This commit marks progress on ACUMEN-138.