Delay discrete assignments to end of discrete step in ref. interp.
This was done by adding tracking of discrete assignments to the
Eval monad. The evalDiscreteAction now adds a triple representing
the assignment computation to the monad instead of applying it
directly. This effectively delays the corresponding state update
until all valuations involved in performing that discrete step
have completed.

The collected assignments are applied in the step function, just
before reparentings are applied.

This relates to ACUMEN-242 (part 1 of semantics revisions).