Implement basic Hybrid Attominton extraction algorithm.
Mostly follows algo.txt in https://bitbucket.org/effective/paper-esaes.

Work in progress: It currently requires a single class model with lots
of other restrictions that are not checked for.

To run my itself run acumen with: sbt "run <model> extract"

To run as a prepass use the new "--extract-ha" option in
GraphicalMain.

The extraction is a first step in getting arbitrary models to run with
the enclosure interpreter.

Towards ACUMEN-257.