Test that examples produce the expected result in the test suite.
That is, for CStore based interpreters and in addition to testing that
all the examples run without error, perform a simple regression test
by compare the results against the expected result.  The expected
result here is a filtered store at 5 points during the simulation at a
time when the results are known to be correct.  The 5 points are the
final result and the results after the first 4 events.

Results are updated by running sbt "run . examples"; the resulting
files should be checked into the repo.

Note that the test suite will ignore any examples that don't have a
result file after issuing a warning.