Change nextStepType to the more direct resultType.
There are now three (instead of two) result types:
  @Discrete @FixedPoint and @Continuous

If the result type is @FixedPoint then the next step will be
continuous.  If the result type is @Continuous or @Discrete the next
step will be a discrete one with a result type of either @Discrete or
@FixedPoint.  Thus, it is possible to determine the nextStepType from
the resultType and hence nextStepType is no longer needed.

The main reason for this change is that resultType is more direct.  It
makes filtering the output based on the step type easier and is less
confusing when viewing the output.