Refactor the imperative interpreters.
The parallel interpreter is the same as the sequential one except for
a parallel traversal.  The refactoring makes this clear and avoids any
duplicate code.

The sequential and parallel subpackages are gone.  The sequential
interpreter is renamed to ImperativeInterpreter and the parallel one
to ParallelInterpreter.  The ParallelInterpreter now extends
ImperativeInterpreter rather than both interpreters extending a
common abstract base class.