Implement basic row and column filtering of the history.
The row filtering is controlled by two new simulator parameters:
outputRows and continuousSkip.  The first controls the type of rows to
output.  The possible values are "All", "WhenChanges" (discrete steps
in which something changed), "FinalWhenChanged" (the final discrete
step after an event), "ContinuousOnly", and "Last" (the very last
step).  The second controls what Continuous steps to output.  A value
of 1 skips every other step, 2, every third, etc.  The default is 0,
which doesn't skip outputting any continuous steps.

(Note, that for technical reasons the initial state (after the first
discrete step) is always saved.)

The column filtering will currently filter out most simulator
parameters and internal fields.  Currently it is not controllable by
the user.

The general framework allows for arbitrary control of what rows and
columns are saved.  This framework is used by the regression test to
save time by avoiding unnecessary work when we are only interested the
results after the first three events and the final result.  With less
data to output it is now possible to test all the examples with the
imperative interpreters (previously there were memory problems with
some of the more complex models and they had to be skipped).

A closely related change is the conversion from using the CStore in
the u.i. and pretty printer to using the new GStore.  A GStore is
similar to a CStore with the following two difference: 1) The
container type is a collection.Iterable instead of immutable.Map and
2) the value type is a generic Value[_] instead of a Value[CId].

With the conversion to using a GStore the results of the imperative
interpreter, with minimal conversion, can be passed around in most
places that expected a CStore.  This change, combined with the default
filtering options, allow the FinalTournamentWithPlayers model to run
again with the imperative interpreters.  By avoiding the unnecessary
conversion to a CStore after each step the imperative interpreter now
runs this model over 4 times faster than the reference interpreter (7s
vs 30s).