Make the Plot/TraceModel immutable.
The primary benefit of this is that there is no longer a need for
complicated rules about accessing data on the EDT as the data is now
immutable.

Accomplishing this turned out to be fairly simple.

For the CStore model all that was needed was a specialized
ArrayBuilder that supported taking snapshots.  By taking advantage of
how an ArrayBuilder works and the fact that Scala uses a G.C., no data
copying was needed for the snapshot.

The Enclosure model already return an effectly immutable data
structure, but a hack was needed to convince Scala of this.

Implements ACUMEN-286.