Started work on a new, principled, generator of Progs.
This generator will attempt to get generation right from the start.
Criteria of interesting models, such as having finite dynamics
thoughout the simulation, will be incorporated by deriving the
continuous assignments of each class from a randomized system
of differential equations for which a solution is known to exist.

This relates to ACUMEN-294.