Merge GraphicalMain and Main command line handling.
Also provide "--help" for normal help and "--full-help" to provide
help with the experimental parts.

The "--example" command line option is replaces with the more sane
"--model" option.  The syntax of the command-line main commands
changed so the command is first and the file is second for example
instead of "model.acm last" use "last model.acm".  This provides a
more uniform syntax inline with other multi-use binaries.

Towards ACUMEN-80.