Added generation of conditionals to Prog generator.
This yields hybrid systems becasue, depending on the value of
conditions throughout the simulation, the active set of continuous
assingments will also change.

The conditions are generated based on the range of each variable
during simulation. A subinterval of the range is generated, and
a condition is constructed which is true when the value of the
variable is within this subinterval.

This relates to ACUMEN-294.