Added generation of discrete assignments to Prog generator.
The implementation tries to avoid generating duplicate discrete
assignments to the same variable within the same scope (same
method is being used to avoid having multiple equations with the
same LHS). This approach does not, however, deal with the situation
when e.g. two consecutive conditionals whose then-branches contain
assignments to a variable trigger at the same time. One way to deal
with this would be to execute the generated model using the updated
reference semantics (which detects this situation when executing).
When this situation is detected, one could either try to remove the
offending discrete assignment from the generated AST, or throw the
model away all together and genereate a new one. The latter will
probably run into problems, as the probability of generating
problematic models increases with the size of the model.

This relates to ACUMEN-294.