Improved the Prog generator now always produces runnable programs.
Done by filtering the output of the generator using a new function
"simulatesOK". This function attempts to execute the Prog p using
the new reference interpreter. It rejects p if an exception is
thrown during execution, e.g. when a repeated assignment is
detected in the interpreter, or if a possible infinite loop is
detected (> 1000 consecutive discrete steps).

There are two downsides with this approach.

  1. The possibility exists that the generator will not be able to
     generate any runnable models and give up. This can be a
     problem if the generator is used in a composite generator.
  2. The generator executes all Prog instances it genererates,
     which will significantly increase the execution time of any
     test based on this generator.

This relates to ACUMEN-294.