Implemented basic generation of models with multiple classes.
A program with multiple classes can be seen as a hierarchy (DAG).
If class A contains the statement "create B(..)" in its ClassDef,
then A is above B in this hierarchy. The Main class is at the top
of the hierarchy and classes which contain no create statements
are referred to as leaf classes.

The basic Prog generation algorithm for multi-class models works
approximately as follows:

  1. Generate list of leaf classes and add these to a environment.
  2. A small, random number of times, repeat the following:
     generate a list of classes, based on the environment, and add
     these to the environment.
  3. Generate a main class based on the environment and wrap it
     in a Prog.

Summary of other changes:

  - Changed to using Dot instead of Name throughout the generators
    to make object field references possible in expressions.
  - Moved utility generator combinators to Generators.scala.
  - The ClassDef._types map is now used to store information about
    the types of generated class fields, used when generating
    object private variable references.
  - Added parameters to ProgGenerator to make it more easily
    reusable in different contexts.
  - Added property for genLinearODE.

This relates to ACUMEN-294.