Attempt to eliminate state variables in the H.A. Extraction.
This is done by adding pre-conds to the modes with the aim of using
that information to eliminate the state variables.

A simple analysis is done to determine what mode to go into after each
reset, if we can't go into a mode a new discrete mode is created with
the pre-conds being the same as the post-conds for the reset.

After all the resets are added a simple analysis is done to prune
unneeded conditionals from each reset based on the modes pre-conds.
If the reset conds. can be shown to be false, than the reset is simply
eliminated

After the resets are pruned kill any unused (discrete) variables with
the assumption that they are only used to determine state.

Other minor improvements to support this.

(Parts of this code is copied from the orignal extract algo.)

Towards ACUMEN-257.