Rework how conditionals are handled.
Conditionals are no longer represented by a list of "Cond"'s, but
rather a Cond itself using the Cond.And type.

matchConds is no more and instead the Cond#eval method is used.  This
spreads the logic out among Cond children and hopefully makes it
easier to understand.

Towards ACUMEN-257.