Change how ensureMode works on matching transient modes.
Before it would attempt to find a transient modes based on evaluating
the mode preconds against the resets postconds.  But that would always
fail due to the the presence of the mode var.  Now instead filter the
pre/post conds to only include conditions corresponding to a likely
state transition and simple search for a mode with a post-cond that
matches the reset pre-cond and if one is not found create a new
transient mode.

Also filter non-transient modes preconds in a similar way, that is to
only include conditions corresponding to a likely state transitions
rather than just discrete variables.

Towards ACUMEN-257.