Use Set rather than Seq in Cond.And.
The main reason for this is so that Cond.And (and thus all Cond's) can
be compared for equality reliably.  Also Set is a logical data
structure to use.  The only reason not to use it was because it might
be inefficient for small sets, but it turns out Scala has specialized
implementation for such small sets.

Towards ACUMEN-257.