Redo the H.A. Extraction strategy; it's now close to the orignal ver.
It turns out that the strategy before this commit was wrong in subtle
ways due to invalid assumptions.  These assumptions become obvious
when I applied more aggressive optimizations.

The new strategy uses many of the ideas from the old (and hence most
of the same code).  The key difference is that the new strategy uses
the simplest possible conversion to a H.A. (in which every leaf in the
if-tree is a mode) as the bases to perform additional optimizations,
rather than trying to use a more complicated conversion algorithm that
minimized the number of modes.  The more complicated conversions ended
being both harder to get right and harder to reason about and was thus
the cause of many of the previously mentioned invalid assumptions.

With an easier to reason about initial H.A., I could apply more
aggressive optimizations and get the new strategy mostly back to the
point where that the original "full" version was.  To fully get back
to that point, this version will need to do something meaningful with
the claims.  This version, could also use better error checking when
invalid modes are passed in.

Towards ACUMEN-257.